pyproject.toml
[tool.poetry]
name = "haybox-cli"
version = "0.1.0"
description = ""
authors = ["Jonathan Haylett <jonathan@haylett.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
cobs = "^1.2.0"
protobuf = "^4.24.3"
click = "^8.1.7"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.58.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
